get_predictions/requirements.txt
absl-py==0.9.0
astor==0.8.1
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2019.11.28
cffi==1.14.0
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.2
Distance==0.1.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
opencv-python==4.2.0.34
html5lib==0.9999999
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
Pillow==7.0.0
urllib3==1.22
PyYAML==3.12
matplotlib==3.2.0
natsort==7.0.1
numpy==1.18.2
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.3
pycairo==1.19.1
pycparser==2.20
pytesseract==0.3.3
PyGObject==3.36.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tensorboard==1.14.0
tensorflow==1.5.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
tesseract==0.1.3
tesserocr==2.5.1
